requirements.txt
celery
python-logstash
